Some changes, added cool changelog, added fonts, changed some HUD shit.

+++ src/main/java/net/minecraft/client/gui/FontRenderer.java
@@ -374,7 +374,8 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string with a shadow.
      */
-    public int drawStringWithShadow(String text, float x, float y, int color)
+
+    public float drawStringWithShadow(String text, float x, float y, int color)
     {
         return this.drawString(text, x, y, color, true);
     }
@@ -382,12 +383,12 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, int x, int y, int color)
+    public float drawString(String text, int x, int y, int color)
     {
         return this.drawString(text, (float)x, (float)y, color, false);
     }
 
-    public int drawString(String text, float x, float y, int color)
+    public float drawString(String text, float x, float y, int color)
     {
         return this.drawString(text, x, y, color, false);
     }
@@ -395,7 +396,8 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, float x, float y, int color, boolean dropShadow)
+
+    public float drawString(String text, float x, float y, int color, boolean dropShadow)
     {
         this.enableAlpha();
